fix(module): PY2 fail bc using same name for package/module
Trying to work across PY-versions is easier if not shadowing
package-name with module, and using relative imports instead.

So move code:
  omniparse.omniparse.py --> omniparse.__init__.py
- and move project coords --> omniparse._version.py.
