+++ omnihash/_version.py
@@ -1,0 +1,5 @@
+__version__ = '0.12.1'
+__license__   = "MIT License"
+__title__     = "omnihash"
+__summary__ = "Hash files/strings/streams/network-resources simultaneously in various algorithms."
+__uri__ = "https://github.com/Miserlou/omnihash"
